+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,13 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.1] - Unreleased
+
+### Changed
+
+- With `portable` enabled, the `set` built-in now rejects a `-` used as a separator between options and operands (for example, `set - foo` or `set -a - foo`) as an error.
+- With `portable` enabled, the shell command line now rejects `-` and `--` given together as option-operand separators (for example, `yash3 - -- myscript`) as an error.
+
 ## [3.4.0] - 2026-08-22
 
 This version was accidentally published as 3.3.5, which was then yanked.
@@ -459,6 +466,7 @@ later.
 
 - Initial release of the shell
 
+[3.4.1]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.4.1
 [3.4.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.4.0
 [3.3.4]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.3.4
 [3.3.3]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.3.3